requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
fastapi==0.115.5
filelock==3.16.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-multipart==0.0.18
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
snowflake-connector-python==3.12.3
sortedcontainers==2.4.0
starlette==0.41.3
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
